Install dotenv and pg
