+++ server/package.json
@@ -11,10 +11,13 @@
   "dependencies": {
     "axios": "^0.20.0",
     "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "faker": "^5.1.0",
     "mysql2": "^2.2.5",
     "nodemon": "^2.0.5",
+    "pg": "^8.4.1",
+    "pg-hstore": "^2.3.3",
     "sequelize": "^6.3.5"
   },
   "devDependencies": {
